fix: allow concurrent rpc calls & de-duplicate provider requests (#60)
Don't double-fetch the same providers request: we are already using this technique to ensure we don't double-fetch bitswap requests.

Rather than storing provider request response channels based on the request Id, store a vector of response channels based on the key. When the `Event::Kademlia(KademliaEvent::OutboundQueryCompleted` event occurs, send the list of providers to each channel that was stored.