feat: `iroh-ctl status -w` watches the health of each iroh process (#84)
Adds `check` and `watch` methods for the rpc `Client`, that return a `StatusTable` or a stream of `StatusTable`s.

A `StatusTable` is a struct with a set of `StatusRow`s, one for the `gateway`, `p2p`, and `store` rpc servers. This is currently hardcoded as having one "slot" for each process, but in the future should be made more flexible to allow for mulitple `gateway`s, `p2p` nodes, and `store`s.

The status of each row is based on `tonic_health::ServingStatus`, and is received by calling `check` or `watch` on the `HealthServer` of each process. Each rpc server now starts up a `HealthServer` service.

In the future, we may want to implement our own version of `HealthReporter` and `HealthServer`, so we can more finely control and pass through the status of each process.