feat(iroh-resolver): create a balanced tree dag
When adding content to iroh, create a balanced tree.

Our balanced tree implementation takes a stream of
chunks (`BytesMut`) and returns a stream of Bytes
of the encoded nodes of the dag, and their associated Cids.
`(Cid, Bytes)`

The implementation emits the node as soon as it has
been created.

All the leaves of the tree are encoded as unixfs `Raw`
nodes. All the stems & the root of the tree are encoded
as unixfs `File` nodes.

We've put effort into having our nodes be as similar to
kubo as possible, but we have intentionally differed in
a few ways:
- all of our "stem" nodes have `name: None`. In kubo, the
parents of leaf nodes have the go equivalent of
`name: Some("".to_string())`.
- the `tsize` of our links is the cumulative size of
the encoded tree associated with that link. The definition
of `tsize` seems to be unfixed, and can mean different things
depending on the implementation.