feat: race a http gateway with p2p in iroh-one's content loader
This adds to iroh-one a new optional configuration item: resolver_gateway
If resolver_gateway is set, the resolver will race the p2p node with
fetching a raw block from the gateway.

A new source is also introduced to classify blocks coming from a gateway out
from those fetched from p2p or store. Hits from the gateway are considered as
not cached for telemetry purposes.