feat: recursive directories & directory link limits (#213)
All adding directories (and recursive directories) to iroh.

Directories with entries that number larger than 6000 should be represented as a HAMT. Attempting to create a directory of that size will error.

We determined the `DIRECTORY_LINK_LIMIT`  based on a 2MB block size, CID size as 64 bytes, name length as max 256 bytes, and tsize size as 8 bytes.  (64 + 256 + 8) / (2 * 1024 * 1024) is about 6400. Leaving a generous buffer, we set our link size to be 6k.