feat: introduce an Api trait and an implementation based around the RPC client
This is exposed as a public API and will also be used to implement the CLI in
iroh-ctl.

The idea is that the trait can be reimplemented to help test the iroh-ctl later,
and it defines clearly what the public API is.

We smooth the underlying client somewhat: make Cid and PeerId into references,
and use &[] slices instead of vectors.