feat: `iroh-ctl add` & `iroh-ctl get` (#164)
implements `iroh-ctl add`, with the `--recursive` flag option for adding directories, and the `--no-wrap` option to NOT wrap the root.

We have defaulted to wrapping the root of the file or directory with a `UnixfsNode::Directory`, so we can preserve the file or directory name. In other words, if you choose the `--no-wrap` option, the name of the file or directory you have added will not be preserved.

also implemented `iroh-ctl get`. The top-level file or directory name will be the cid, unless you use the `--output` flag, which will replace that cid with the output path.