feat: iroh-api rename, iroh get basics, rpc client construction
## move RPC client construction into `iroh` (read: `iroh-api`)

I've moved creation of the RpcClient away from the CLI into `iroh` (to be `iroh-api`). This means that the API gains a bunch of dependencies and the CLI loses them.

Doing this made me realize that the api should own the rpc client, so changed from a reference (with a lifetime) to ownership.

## `iroh get` start of implementation

Start of implementation of `iroh get` with a few mocked CLI tests

## `iroh` -> `iroh-api` crate

The plan is to have an `iroh` crate that contains the CLI and an `iroh-api` crate that contains the Rust API. Here we've done half of that renaming: `iroh` is renamed to `iroh-api`. I haven't done the second half, renaming `iroh-ctl` to `iroh`, as that would make this commit too different to read.