feat: CLI tests for iroh add, and API changes (#343)
This contains work on the add command, with a few CLI tests and some API tweaks.

## add layering

`add_file` and `add_dir` in the core `Api` that's mockable, and the slightly higher level `add` is above it in `ApiExt`.

## recursive simplication

Test for the `-r` flag in the CLI and just let `add_dir` in the resolver add a directory without worrying. This is tested.

## a few add CLI tests

A few basic add CLI tests.

## CLI test cleanup

I've went through the `cli_tests.rs` module step by step cleaning and reorganizing the tests so it's more easy to maintain them. Unfortunately `trycmd` gets too finicky if I try to use its globbing feature (as input and output directories seem shared sometimes). So now I maintain an explicit list in the same order as they show up alphabetically. It's too verbose but it's the best we can do for now.