feat: progress bar for iroh add command (#368)
This introduces a progress bar for the `iroh add` command. It touches the following layers:

* the various `add_dir`, `add_file` functions in `iroh::resolver::unixfs_builder` now return a stream of `AddEvents`

* these are exposed in the `iroh-api` where they are mockable. The `iroh-api` also provides a higher level method to resolve a `AddEvent` stream to a CID

* `iroh::size` introduces a way to calculate total size of each file on the filesystem

* The `iroh add` command in the `iroh` crate then calculates the total size with a spinner, and then uses this to show a progress bar.

Note that `indicatif`, the progress bar library, allows a lot of tweaking of what we see, both color, characters and text. I made a selection to get started. Suggestions for tweaks of what the bar looks like are welcome.

I want to also include current file information in the message part of the progress bar, but getting that information requires more rework of the `unixfs_builder` and I think we should punt on that for the time being.