ref(lru-cache): Change lru cache crate to avoid vulnerable dependency (#478)
* perf(lru cache): Benchmarks for LRU caches

Some quick and dirty benchmarks for the caches and lru crates for the
LRU cache.

* Improve benchmarks: measure just the operation

This makes use the the facilities to compare two functions and also
uses iter_batches to handle setup and dropping correctly.  The results
are widely different!  The differences between both implementations
are much muchness really, not much to go either way.

* feat(lru): Switch implementation to lru crate

The performance is pretty similar and we don't use any features of the
bigger crate.

* Use current version of criterion

* Remove caches crate entirely

This makes sure it no longer shows up in cargo-audit.

* remove old crate entirely