fix: Correct dependencies and build checks
Fixes some CI commands to check all code by enabling all features and
checking more parts of the code.

As a result fixes the dependencies of the uds-gateway feature and
switches it from the deprecated tempdir to tempfile.

Makes the version spec of tempfile consistent between all the
workspace crates.

Fixes a clippy warning about cloning a bool, which is Copy.

Finally a few extra changes in ci.yml which the linter was complaining
about.