chore(ci): Test more features and MSRV
Adding explicit check and clippy steps for the default features,
because --all-features enables the "testing" feature which in turn
disables a bunch of production code.  Preferably we figure out how to
not make --all-features skip prod code like this, but this is a stop
gap in the mean time.

Adding an explicit MSRV check and removing the rust-toolchain file.
This means local dev will default to using the latest rust compiler
and we'll not be surprised by failing CI that worked locally.  The
explicit MSRV check makes it so that we at least notice when we break
this.