Fix QUIC stream completion (#143)
Sending a STOP_STREAM frame from the receiver is the way for a receiver
to go to the reset state. It involves sending an extra frame, the
receiver will then send its own RESET_STREAM frame and the stream is
considered ended in an error state.

Instead the right way to end a stream is to set the FIN bit on the final
STREAM frame, ideally not even resulting in an extra frame or packet
sent. This is done on the sender side.