feat: Remove MAX_DATA_LIMIT
We used to limit the maximum amount of data that can be transferred,
both the total size as well as single blob size.  There is no real
benefit to this, we pass through the reader and do not buffer all this
data at once.  And avoiding to fill up the user's disk space is not
our concern.