fix: Improve listening addr output
This fixes the log message output which had an extraneous newline in
it and needlessly used debug output.

It also adds the listening address to the normal provider output: This
is universally useful since you can specify addresses with -a but
don't otherwise know which one is used if using OS-allocated ports.
This can also be useful for testing.