ref: Remove duplicate read_lp implementation
Seems like we ended up with two bits of code doing exactly the same.
Clean it up.

This settles on read_lp() as name since this is symmetric with
write_lp() which writes bytes.  This implementation inlines the
functions a bit more as well, the entire function is still small and
it is used often.