feat: begin impl Server side of derp, starting with the server side of the client connection (#826)
* added `Conn` trait, `ClientConn`, & added more to `Server`, 
* `ClientConnManager`, `ClientConnWriter`, `ClientConnReader`
Implement structures that manage the connection from the server to the
client:

When the server "accepts" a connection (`accept` is not yet
implemented), it separates the write and read portions of the
connection. It should create a `ClientConnManager` to interact with the
connection. The `ClientConnManager` runs the `ClientConnWriter` and `ClientConnReader` in a loop.
The `ClientConnWriter` accepts notifications & relayed packets from the server & sends them to the client.
The `ClientConnReader` to reads any notifications or packets it wants
the server to relay from the client, & sends them to be processed in the
server.
The server should hold onto the `ClientConnManager`, and use that as
a handle to communicate with the `ClientConnWriter` to forward any
relayed packets, or to shutdown the connection with the client if
something goes wrong.