add `Client` and `Clients`
`Client` will eventually be able to handle multiple `ClientConnManager`s, but for now, the lastest connection will be the one we hold onto

The server will use the `Clients` struct to manages all the clients connection to the server. It is how it will actually send the packets from one client to another.