feat: Add run_ticket to dial all addresses stored in a Ticket
This adds a new iroh::get::run_ticket call which will dial all the
addresses in a ticket concurrently, up to a specified concurrency
limit.  The first established connection is returned.

This allows much better fetching of a ticket, without having to know
how it is constructed and speeds up connecting.

When dialing an attempt is made to first try addresses which are
considered to be on the same subnet as any local address.