feat(ticket): Ensure a ticket always has at least one address (#892)
This makes it impossible to have a ticket with an empty address list.
It makes the fields of the ticket private and also makes creating the
ticket fallible.  Both are probably good things for the public API and
the latter e.g. allows us to add further checks like ensure the hash
really exists.