Minimal fix for the disco bug
We now add a length prefix to all packets. We also added this length prefix
to disco packets, meaning that they were no longer disco packets.

In principle I think this is a good thing - you can now transmit any byte
string, including the disco magic string, as a normal packet without it
being accidentally recognized as a disco packet.

But a proper fix for this would involve looking through the code base for
places where the magic byte is used and replacing these with a dedicated
path for disco messages. I tried it, but wasn't sure I got them all.