ref(netcheck): statically declare udp bind addresses
This is just a PR to keep them smaller.  The main chainge is that the
UDP bind sockets to do STUN on are now directly statically declared.
Functionally this is exactly the same but without the extra state and
function calls involved.

Also adds some doc comments and mostly syntactic refactoring.