Drop conn (#1029)
* feat: Add drop for conn

this makes sure we close a connection on drop.

* ref: remove provider shutdown

this is now taken care of by the conn drop

* feat: make conn drop a bit more drastic

this should be like kill -9. We don't want to wait until the conn actor
processes the shutdown msg, but kill it with prejudice.

If somebody wants a more gentle shutdown they should use shutdown. That's
what it is for.

* fix: owned join handle

whenever this gets dropped, the task gets aborted. Which means that the
task will be dropped as soon as it yields (which it should do since it is
a normal tokio task)

* pr: rename to AbortingJoinHandle