ref(netcheck): make netcheck a long-running actor (#1028)
This turns the netcheck into a long running actor:

    The Client creates the actor which runs in a background tokio task.
    The Client can be cloned to get more handles to the netcheck actor.
    Once Client is dropped the actor is stopped.
    Now you can always pass stun packets to to netcheck: it manages the state internally.

This last point is the motivation, it simplifies how things are wired up and makes it easier to keep the right state in the netcheck actor where it belongs. This will also make future changes to netcheck more contained.

It turns the former netcheck actor into more of a full-scale actor. The ReportState is not following this pattern yet.