ref(metrics): Document metrics and remove macro_use (#1045)
This improves some documentation by explaining the different metric
types and the difference between record and observe.  It's also a bit
opinionated about metric types.

It also restructures the code a bit to make use of the 2018-edition
macro scoping rules, so macros are not magically available and use can
import them from the metrics module, which gives much better context.