tests: Add test utilities to configure logging
This is useful in many tests, not just netcheck.  Make it more
globally available.

It also makes it better to enable by default: it just always logs
all modules on the TRACE level.  But it is properly captured by the
test runner and only shown if the test fails.

Potentially this is too verbose for some tests.  Variants can be built
that allow configuring the logging level or even per-modules logging
level as required.  But no need for that just yet.