tests: Add test utilities to configure logging (#1060)
This is useful in many tests, not just netcheck.  Make it more
globally available.

It also makes it better to enable by default: it just always logs
all modules on the TRACE level.  But it is properly captured by the
test runner and only shown if the test fails.

To change the logging the RUST_LOG environment variable can
be used like normal to override the TRACE logging.