fix: increase broadcast channel size
before, event producer and event consumer were on the same thread of the
same runtime. Now the events are produced on one of the tpc threads, and
the consumer is the generic tokio runtime. So we need a larger buffer.
Especially the many files test can produce a lot of events in a short time.