Feat: Add iroh runtime wrapper and actually use thread per core (#1088)
* feat: add our own runtime wrapper and use it

* fix: increase broadcast channel size

before, event producer and event consumer were on the same thread of the
same runtime. Now the events are produced on one of the tpc threads, and
the consumer is the generic tokio runtime. So we need a larger buffer.
Especially the many files test can produce a lot of events in a short time.

* ref: simplify type constraints for tpc runtime handle

* test: add a cli test that transfers a large file

just a single file to have a rough idea about perf. Run with

cargo test --release cli_provide_one_file_large -- --ignored --nocapture

* fix: switch runtime back to multithreaded

* refactor: simplify dropping/shutdown of tpc runtime

and also reduce the number of threads of the main/misc runtime to 2, since
we do not need that many. The main work is done in the tpc runtimes.

* pr comments:

- add expect instead of unwrap
- add from_current utility
- rename _sync to _blocking

* docs: add more comments, in particular about panic behaviour

* ref: make most of the spawn methods test-only

also require futures to be boxed for tpc runtime. The reason for this is
that if the futures are already boxed for some reason we don't want to box
twice.

* pr: add spin_loop hint