ref: Change module structure and visibility and require docs
This modifies the modules inside the crate structure and visibility.
Specifically it moves the `config` and `command` modules from the iroh
*crate* into the iroh *binary*.  This could be a little confusing.
This means they are no longer part of the public API of the iroh
crate (and do not need to be documented).

It also fills in a few odd bits of missing doc comments and makes doc
comments required for public APIs on the crate.