ref(netcheck): Turn ReportState into more of an actor (#1103)
This moves ReportState into an "Actor" like netcheck is. It still runs most of the report tasks as futures (haircheck has become an actor, it receives messages). The main effect of this is that ReportState (now reportgen) does not have any locks around the report anymore. The control flow does not change significantly (yet) but the amount of state passed around and the amount of local state needing to be stored has reduced significantly.

Closes #1100
Closes #1037
Closes #1012