Simplify iroh-bytes get (#1202)
* feature: add fn for getting the initial state of the get response machine

* refactor: remove lots of code from iroh-bytes::get

this is all not really related to the iroh-bytes protocol but only relevant
for iroh itself. So in order to keep the surface area small, let's move it
into iroh.

* refactor: move these methods into iroh::util for now

* refactor: reorg the various utils that we sadly have to have

* refactor: remove unneeded reexport

* refactor: more moving utils around

* refactor: rename `get_response_machine` to `fsm`.

Fsm is used also in bao-tree, so I find it is justified to introduce this
abbreviation. The module used to be more top level, so `get_response_machine`
was more justified. Now that it is in `get` I think `get_response_...` is a
bit too verbose.