ref(logging): Use nested tracing spans (#1228)
This does a fair amount of logging-only changes.  Mostly it hooks up
the tracing spans to nest correctly.  This results in seeing all the
outer contexts when reading messages (at the cost of shifting plain
output to the right).

As a side effect logging the Conn name is not needed on many methods
since it is already present in the parent span.  So this removes a lot
of duplicate logging.