fix(netcheck): Build test ProbePlan from fake interface data
The tests are supposed to be purely algorithmic, but they were reliant
on the local interfaces discovered which made them fickle.  This
spells out the interface which is much more reliable.

More variations in the interface states are probably useful too, but
for now this gives us some basic coverage of probe plans.

Closes #1263