fix(iroh-net): Stop quinn from erroring when closing a connection
Quinn spawns a driver task for each endpoint.  This task only
terminates when the endpoint is completely dropped.  Crucially however
at this point it will still poll the underlying socket once more and
if the socket errors at this time it will report the error.  Only when
the driver task terminates does it drop the socket, which must free
it's resources on drop.

This implementation does not do anything to make the MagicSocket shut
down cleanly on drop: dropping it drops the AbortingTaskHandles, in
turn aborting the associated tasks of the MagicSocket.