fix(iroh-net): Split GSO packts received from derper
The datagrams received over the derper channel could be multiple
datagrams combined by Generic Segmentation Offloading (GSO).  Normally
these would go over the network as individual datagrams and the
receiving machine would use GRO if it had it and all would look normal
for the application code reading from the sockets.

However when sending datagrams over the derper channel they arrive at
the receiver in whichever way the GSO of the sender assembled them.
We could stop that from happening, but we might as well use GSO/GRO
over the derper channel so we benefit from it there too.

So instead this manually implements partial GRO on the receiver side.
It splits up the received datagrams into small enough chunks so they
can be received by the host's IoSlices (quinn configures these smaller
depending on the host's limit so save wasting memory).  This can still
read multiple datagrams in as one packet, but the RecvMeta's stride is
set correctly and Quinn handles this fine.  So it delays splitting
into individual datagrams as long as possible.

Note that this is kind of a hot-fix.  We might want to optimise
GSO/GRO paths through magicsock a bit more and this could end up being
refactored in different ways.  See also #1358.

Fixes #1357