fix(iroh-net): split packets on send (#1380)
this seems to be the safest way to fix possible GSO/GRO issues on all
platforms (Linux and ???) that support them.

## Description

This is a fix for the possible issues described in
https://github.com/n0-computer/iroh/issues/1357 that does the simplest
possible thing - just split up the packets before sending them to the
derper.

Note that we still get GRO/GSO when we have a direct connection. And
since the stream to the derper is a byte stream, unless we flush after
each send this should still be rather efficient. The only downside is
that the derper has to deal with more individual packets.

## Notes & open questions

Does this fix the issue?

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.