refactor(iroh-net): use tokio-util::codec for derp protocol implementation (#1386)
We have been having issues multiple times with the buffer management,
this cleans up things and should make them more reliable. The core
change is to use `tokio_util::codec` to implement the derp protocol.
Additionally there are a couple smaller fixes in the derper code and
tests, for issues that where discovered while doing this.

Fixes #1384, #1381, #1247

---------

Co-authored-by: Ruediger Klaehn <rklaehn@protonmail.com>
Co-authored-by: Kasey <klhuizinga@gmail.com>