fix(sync): properly drop rpc subscriptions, add doc status (#1396)
## Description

So far in #1333, if a RPC or in-memory client called `doc.subscribe()`
the event callback would never be dropped, even if the client did drop
the event stream. This PR fixes this, by having the event callbacks
return whether the callback should stay active or not. We can't use the
removal token here, because calling `LiveSync::unsubscribe` from within
the event callback would deadlock the actor.

Also adds a a `LiveStatus` to the doc info RPC call. For now only
contains the number of subscribers. More info, e.g. on peers, can come
later.

## Notes & open questions

* As of #1333 and unchanged by this PR: `doc.subscribe` will fail for
documents that are not in the `LiveSync` (they are added via
`doc.import` or `doc.start_sync`). This is unfortunate, because you'd
often want to setup a subscription before starting sync, to catch all
events.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.