test: Introduce iroh-test with common logging infrastructure (#1365)
## Description

This code was copy-pasted around too much.  This takes a simple and
clean approach to this by creating a new crate which is used as a
dev-dependency of all the other crates.

The upside is there is no messing about with cargo features and
delicate matching versions.

The main downside is that it can not provide testing utilities which
build on the code in the other crates.  Most notable this is currently
the testing derp server.


## Notes & open questions

This skirts around the tricky bits of adding such a common module: The
testing derp server is still copy-pasted around.  I think this is
fine, because solving that is ugly.  Let's see if any other code ends
up in this crate or if the majority needs a more complicated solution
anyway.

This also doesn't introduce a general *published* iroh-common crate
(or whatever it would be named).  This might also be useful but
affects the public API and introduces other questions.  There's no
harm in considering this to be an entirely independent problem and to
later add such a module as well as this one.


## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.