feat(iroh-bytes): remove unneeded u64 length prefix (#1408)
## Description

we used to write the length prefix here. But this is not needed since we
immediately close the stream after writing the request. And since this
is a request/response protocol this is unlikely to change.

quinn::ReadStream::read_to_end offers a size limit where we can just use
MAX_MESSAGE_SIZE to be on the safe side.

## Notes & open questions

Can somebody think of a scenario where the length prefix is useful?

## Change checklist

- [x] Self-review.