refactor: Reduce amount of info logging (#1418)
## Description

There is a bunch of netcheck stuff that happens periodically and does
not need info logging for this.

Similarly this reduces the log level of some very regular logging in
the derp client, magicsock and iroh.

Lastly log_concise_report is updated to not do so many allocations
while building the log message.


## Notes & open questions

This does not yet (knowlying) address the issue of the ios debug
console becoming unusable due to logging.

## Change checklist

- [x] Self-review.
- [x] ~~Documentation updates if relevant.~~
- [x] ~~Tests if relevant.~~