feat(iroh-net): add `DEV_DERP_ONLY` env variable for testing the derp relay (#1378)
## Description
This forces the `MagicSock` to only ever *send* messages through a derp
node.

However, this does not prevent the node from receiving messages over the
udp sockets.

Hole-punching attempts and responses are prevented. 

------

Adjusts the `addr_for_send` method, ensuring that we only ever return
the `DerpRegion` (if we know it) as a possible option for how to
communicate with the given peer.

It also prevents `send_pings` and `start_pings`.

Both the provider and getter must use the `DEV_DERP_ONLY` flag for this
to function. There is some bad interaction that causes blocks / stalling
when we attempt to hole punch and use the relay.

-----

Fixes a bug around discovering derp regions. Now, when we receive a
message from a `PublicKey` from a particular `DerpRegion`, and we
previously had no derp information about this `PublicKey` before, we add
that region ID to the `Endpoint` associated with that `PublicKey`.

Also fixes some tests that previously ignored adding derp information.

Now, if you run `DEV_DERP_ONLY=true cargo test`, all tests will pass
with any connections strictly passing through the derp relay.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.