No limits (#1429)
## Description

As discussed, remove limits for now. Also get rid of the ability to
define a custom contains in RangerKey for safety.

## Notes & open questions

I just mechanically went through the code and removed all limits. It
might be that there are some structs that only make sense in the
presence of limits that should also be removed.

## Change checklist

- [x] Self-review.
- [x] Tests if relevant.