refactor(iroh-sync): process items one by one without preprocessing
Discussed this with AljoscheMeyer on discord. I think in the long term I want
to rely on these things being sorted, but for now this just processes the
ranges individually.

"That depends on how strict do you define your protocol. One approach is
to require no such things and simply process arbitrary ranges individually
as they come in. The other extreme is to consider overlapping or unordered
ranges as invalid. I don't believe that's a good idea, as it adds conceptual
complexity.