refactor(iroh-sync): store `RecordIdentifier` as bytes (#1448)
Builds on #1445. Had an idea - we can make the `RecordIdentifier` be a
newtype wrapper around `Bytes`. With this we can expose the
`RecordIdentifier` directly as a simple continous byte slice.

I'm not too sure if we gain much over the previous form – Maybe some
performance when cloning oftenly. And @rklaehn wanted to have a simple
byte representation for store improvements.