refactor(iroh-net): remove `NetworkMap` (#1447)
## Description

While having another look at the code when thinking about #1440 I
realized that the `NetworkMap`is never used apart from serving as a
source of entries for the `PeerMap`.

So I removed it. Looks much simpler now, and tests passed for me
locally. Not sure if I haven't missed something.

## Notes & open questions

* We never remove entries from the `peer_map`. This is already the case
before this PR if I'm not missing something, because we never removed
entries from the `NetworkMap`, and only the latter would ever have
triggered a removal of entries from the `peer_map`.

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.