test(iroh-net): Removed unused stun_test_ip field from DerpNode (#1450)
## Description

The DerpNode had a field to override the IP address of the STUN server
in testing scenarios.  However we do not use this, it is always the
same address as the derp server, just like in production.

Removing this removes some code complexity.

## Notes & open questions

I believe this was an unused left-over from the wireguard conversion.

## Change checklist

- [x] Self-review.
- [x] ~~Documentation updates if relevant.~~
- [x] Tests if relevant.