(iroh-net): remove our public key from the endpoint (#1482)
## Description

While trying to understand the information we store for peers for #1481
I noticed this can be removed. It's more efficient in space in a couple
of ways:
- It's not duplicated for every known peer
- Does not need to be cloned for every sent ping
- The `ActorMessage::SendDiscoMessage` that was removed had a size of
512bytes, making it the largest message of the enum. By replacing it
with the two new ones the size of the `ActorMessage` size is reduced
from 512 to 248. The whole change reduces the size of `Endpoint` from
784 to 560.

## Notes & open questions

One place where this might be potentially less efficient is in calling
`public_key()` on the `secret_key` for every sent ping. If this
computation is expensive we can just cache the public key and use it
there. Maybe @dignifiedquire knows if this should be done or if it's
fine like this

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.